Use abstract class TaskManagerTest for testing managers
